Add Samsung print recommendation plugin.
The code follows the same pattern as the HP and Mopria plugins. As this code is contributed by a 3rd I'd rather submit this as is and let the original submitter clean it up once it appears in AOSP.

The code should not affect anything outside of the package it is in. In the package the code is strongly isolated so that one plugin cannot cause issues with another plugin.

The code was minimally tested on a Samsung SCX-3406FW .

Bug: 28908572
Change-Id: I33c2673d06711902dbd0f1d82563cf5d3fbfe86a
(cherry picked from commit 51bd48603a2fc1f34976ae08e9264d496daf60aa)