Fixed a few memory leaks in SystemUI
Fist of all we made the statemap a weak hashmap in order never
to leak any views anymore.

Another leak could occur because predrawlisteners were added twice
but removal only removes one. A view was then leaked in case it was detached
before predraw.

Also fixed a leak when transforming the header because a state wasn't
nulled.

Change-Id: I2573a506c307196ef60c905dc823ea8a95e91a16
Fixes: 28945863