Fixes issue with group backgrounds when animating to keyguard
The backgrounds weren't being updated before the collapse animation when
going to the keyguard so it would look very bad when they collapsed.

This moves updating the background into setChildrenExpanded which will
cover updating the background for both the tapping to expand case as
well as the animating to the keyguard case.

Change-Id: Icba2fa895498d307a38bd9f65be06265afd21a1a
Fixes: 29240532