Fix Bluetooth OOB pairing not working for unseen devices
When Out Of Band info is provided, first time we learn about
device is when it's state is changed to "Bonding". This currently cause
this event to be discarded and pairing fails. Instead, we should add
CachedDevice object for newly added device.

Bug: 30188629
Change-Id: I4046bfe30e8bf32b888f608ac06d5041e2999dc9
(cherry picked from commit 662f333dfbe3117e0092bcc05ba3900443755657)